# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Fingarden API Docs",
    "version": "1.0.0",
    "title": "Fingarden"
  },
  "basePath": "/",
  "paths": {
...
    "/wishlists": {
      "get": {
        "tags": ["wishlist"],
        "summary": "Get array of wishlists",
        "description": "",
        "operationId": "getWishlists",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sorting wishlists by property",
            "default": "productId",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordering results. ASC | DESC",
            "default": "DESC",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limiting wishlists in array",
            "default": 10,
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Wishlist"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["wishlist"],
        "summary": "Create a wishlist",
        "description": "",
        "operationId": "createWishlist",
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "successful created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Wishlist request"
          }
        }
      }
    },
    "/wishlists/{wishlistId}": {
      "get": {
        "tags": ["wishlist"],
        "summary": "Find wishlist by id",
        "operationId": "findWishlistById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "wishlistId",
            "in": "path",
            "description": "ID of wishlist to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Wishlist"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Wishlist not found"
          }
        }
      },
      "put": {
        "tags": ["wishlist"],
        "summary": "Update the wishlist",
        "description": "",
        "operationId": "updateWishlist",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "wishlistId",
            "in": "path",
            "description": "ID of wishlist to return",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Wishlist"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful updated",
            "schema": {
              "type": "object",
              "properties": {
                "raw": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": ""
                },
                "affected": {
                  "type": "integer",
                  "description": "qty of updated raws"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Wishlist request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Wishlist not found"
          }
        }
      },
      "delete": {
        "tags": ["wishlist"],
        "summary": "Delete the wishlist",
        "description": "",
        "operationId": "deleteWishlist",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "wishlistId",
            "in": "path",
            "description": "ID of wishlist to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deleted",
            "schema": {
              "type": "object",
              "properties": {
                "raw": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": ""
                },
                "affected": {
                  "type": "integer",
                  "description": "qty of deleted raws"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Wishlist not found"
          }
        }
      }
    },
    "/wishlists/wishlistProducts/{wishlistId}": {
      "get": {
        "tags": ["wishlist"],
        "summary": "Find products by wishlist id",
        "operationId": "findWishlistProducts",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "wishlistId",
            "in": "path",
            "description": "ID of wishlist to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful updated",
            "schema": {
              "type": "object",
              "properties": {
                "raw": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": ""
                },
                "affected": {
                  "type": "integer",
                  "description": "qty of updated raws"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Wishlist request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Wishlist not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Brand": {
      "required": ["name", "url"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false,
          "description": "value true can be post only with admin token"
        }
      },
      "xml": {
        "name": "Brand"
      }
    },
...
    "Category": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "parent": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "children": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "parameters": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Parameter"
          }
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Color": {
      "required": ["name", "url", "code"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ProductWithoutJoins"
          },
          "readOnly": true
        },
        "url": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      }
    },
...
    "Parameter": {
      "required": ["name"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Rating": {
      "properties": {
        "1": {
          "type": "number"
        },
        "2": {
          "type": "number"
        },
        "3": {
          "type": "number"
        },
        "4": {
          "type": "number"
        },
        "5": {
          "type": "number"
        },
        "avg": {
          "type": "number"
        }
      }
    },
    "ParameterProduct": {
      "properties": {
        "productId": {
          "type": "string"
        },
        "parameterId": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "parameter": {
          "$ref": "#/definitions/Parameter"
        }
      }
    },
...
    "Product": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "shortDesc": {
          "type": "string"
        },
        "keywords": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "category": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "brand": {
          "type": "object",
          "$ref": "#/definitions/Brand"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/TagWithoutJoins"
          }
        },
        "sizes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/SizeWithoutJoins"
          }
        },
        "url": {
          "type": "string"
        },
        "rating": {
          "$ref": "#/definitions/Rating"
        },
        "reviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewWithoutJoins"
          }
        },
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Question"
          }
        },
        "parameterProducts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ParameterProduct"
          }
        },
        "productVariants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariant"
          }
        }
      }
    },
...
    "ProductWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "url": {
          "type": "string"
        }
      }
    },
    "ProductVariant": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "price": {
          "type": "integer",
          "minimum": 1
        },
        "oldPrice": {
          "type": "integer"
        },
        "wholeSalePrice": {
          "type": "integer"
        },
        "available": {
          "type": "boolean"
        },
        "images": {
          "type": "string"
        },
        "color": {
          "type": "object",
          "$ref": "#/definitions/Color"
        }
      }
    },
...
    "TagWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "SizeWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Review": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "rating": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10
        },
        "text": {
          "type": "string"
        },
        "images": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewReaction"
          }
        }
      }
    },
    "Question": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuestionComment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReactionQuestion"
          }
        }
      }
    },
...
    "ReviewWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "rating": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10
        },
        "text": {
          "type": "string"
        },
        "images": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false
        },
        "productId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewReaction"
          }
        }
      }
    },
...
    "ReviewReaction": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "reviewId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "ReactionQuestion": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "questionId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "User": {
      "required": ["firstName", "lastName", "email", "password"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "isVerified": {
          "type": "boolean",
          "default": false
        },
        "role": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "Comment": {
      "properties": {
        "id": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "review": {
          "type": "object",
          "$ref": "#/definitions/Review"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "reactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentReaction"
          }
        }
      }
    },
    "QuestionComment": {
      "properties": {
        "id": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "question": {
          "type": "object",
          "$ref": "#/definitions/Question"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "reactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentReaction"
          }
        }
      }
    },
    "CommentReaction": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "commentId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "Wishlist": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/WishlistProduct"
          }
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
...
    "WishlistProduct": {
      "properties": {
        "id": {
          "type": "string"
        },
        "productId": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        }
      }
    },
...
  }
}